Fix dependency cycle for good by lifting shared utility functions
